requirements.txt
oracledb
questionary
pyserial